alpha bug2/long
git-svn-id: svn+ssh://ci.ruby-lang.org/ruby/branches/ruby_1_3@435 b2dd03c8-39d4-4d8f-98ff-823fe69b080e

+++ bignum.c
@@ -220,7 +220,7 @@ rb_str2inum(str, base)
     }
 
     if (len <= (sizeof(VALUE)*CHAR_BIT)) {
-	unsigned int val = strtoul((char*)str, 0, base);
+	unsigned long val = strtoul((char*)str, 0, base);
 
 	if (POSFIXABLE(val)) {
 	    if (sign) return INT2FIX(val);
